x-pack/filebeat/input/{websocket=>streaming}: rename package
