+++ CHANGELOG-developer.next.asciidoc
@@ -68,6 +68,7 @@ The list below covers the major changes between 7.0.0-rc2 and main only.
 - Split split httpmon out of x-pack/filebeat/input/internal/httplog. {pull}36385[36385]
 - Beats publishing pipeline does not propagate the close signal to its clients any more. It's responsibility of the user to close the pipeline client. {issue}38197[38197] {pull}38556[38556]
 - Debug log entries from the acker (`stateful ack ...` or `stateless ack ...`) removed. {pull}39672[39672]
+- Rename x-pack/filebeat websocket input to streaming. {issue}40264[40264] {pull}40421[40421]
 
 ==== Bugfixes
