+++ filebeat/docs/filebeat-options.asciidoc
@@ -94,7 +94,7 @@ You can configure {beatname_uc} to use the following inputs:
 * <<{beatname_lc}-input-syslog>>
 * <<{beatname_lc}-input-tcp>>
 * <<{beatname_lc}-input-udp>>
-* <<{beatname_lc}-input-websocket>>
+* <<{beatname_lc}-input-streaming>>
 
 include::multiline.asciidoc[]
 
@@ -156,4 +156,4 @@ include::inputs/input-udp.asciidoc[]
 
 include::inputs/input-unix.asciidoc[]
 
-include::../../x-pack/filebeat/docs/inputs/input-websocket.asciidoc[]
+include::../../x-pack/filebeat/docs/inputs/input-streaming.asciidoc[]
