+++ x-pack/filebeat/input/default-inputs/inputs_other.go
@@ -25,7 +25,7 @@ import (
 	"github.com/elastic/beats/v7/x-pack/filebeat/input/netflow"
 	"github.com/elastic/beats/v7/x-pack/filebeat/input/o365audit"
 	"github.com/elastic/beats/v7/x-pack/filebeat/input/salesforce"
-	"github.com/elastic/beats/v7/x-pack/filebeat/input/websocket"
+	"github.com/elastic/beats/v7/x-pack/filebeat/input/streaming"
 	"github.com/elastic/elastic-agent-libs/logp"
 )
 
@@ -44,7 +44,7 @@ func xpackInputs(info beat.Info, log *logp.Logger, store beater.StateStore) []v2
 		awscloudwatch.Plugin(),
 		lumberjack.Plugin(),
 		salesforce.Plugin(log, store),
-		websocket.Plugin(log, store),
+		streaming.Plugin(log, store),
 		netflow.Plugin(log),
 		benchmark.Plugin(),
 	}
